refactor(ui): move bench dialogs to AppLayout like SettingsModal; drop menu hacks
The New/Change Bench dialogs lived inside AppSidebar and were opened from the
dropdown with a DOM-polling workaround. SettingsModal works cleanly because it
is rendered in AppLayout and toggled by a v-model the menu item flips via an
emitted event — the dialog is never a child of the dropdown's portal subtree,
so nothing overlaps or inerts it. Mirror that: extract BenchSwitcherDialog and
NewBenchDialog, render them in AppLayout, have the sidebar just emit
change-bench/new-bench, and delete openAfterMenuCloses and the pointer-events
poking entirely.
