+++ admin/frontend/src/components/NewBenchDialog.vue
@@ -1,0 +1,144 @@
+<script setup>
+import { ref, computed, watch } from 'vue'
+import { Button, Dialog, ErrorMessage, FormControl } from 'frappe-ui'
+
+const props = defineProps({ modelValue: Boolean })
+const emit = defineEmits(['update:modelValue'])
+
+const show = computed({
+  get: () => props.modelValue,
+  set: (val) => emit('update:modelValue', val),
+})
+
+const name = ref('')
+const processManager = ref('systemd')
+const adminDomain = ref('')
+const error = ref('')
+const creating = ref(false)
+const status = ref('')
+
+const processManagerOptions = [
+  { value: 'systemd', label: 'Systemd', hint: 'Recommended' },
+  { value: 'supervisor', label: 'Supervisor', hint: 'Alternative' },
+]
+
+watch(show, (open) => {
+  if (!open) return
+  name.value = ''
+  processManager.value = 'systemd'
+  adminDomain.value = ''
+  error.value = ''
+  creating.value = false
+  status.value = ''
+})
+
+async function waitUntilLive(port, attempt = 0) {
+  try {
+    const response = await fetch(`/api/benches/ready?port=${port}`)
+    if (response.ok) {
+      const data = await response.json()
+      if (data.ready) {
+        status.value = 'Bench created! Redirecting you to setup…'
+        window.location.href = `${window.location.protocol}//${window.location.hostname}:${port}`
+        return
+      }
+    }
+  } catch { }
+  if (attempt >= 60) {
+    error.value = 'New bench admin server did not come up in time.'
+    creating.value = false
+    return
+  }
+  setTimeout(() => waitUntilLive(port, attempt + 1), 1000)
+}
+
+async function createBench() {
+  const benchName = name.value.trim()
+  if (!benchName) return
+  if (!/^[a-zA-Z0-9_-]+$/.test(benchName)) {
+    error.value = "Bench name must contain only letters, numbers, '-' and '_'"
+    return
+  }
+  const domain = adminDomain.value.trim()
+  if (!domain) {
+    error.value = 'Admin domain is required so the bench is reachable.'
+    return
+  }
+  error.value = ''
+  creating.value = true
+  try {
+    const response = await fetch('/api/benches/new', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ name: benchName, process_manager: processManager.value, admin_domain: domain }),
+    })
+    const data = await response.json()
+    if (!response.ok) {
+      error.value = data.error || 'Failed to create bench'
+      creating.value = false
+      return
+    }
+    status.value = 'Bench created — waiting for it to come up…'
+    waitUntilLive(data.port)
+  } catch {
+    error.value = 'Failed to create bench'
+    creating.value = false
+  }
+}
+</script>
+
+<template>
+  <Dialog v-model="show" title="New Bench" size="lg" :showCloseButton="true">
+    <template #default>
+      <div class="flex flex-col gap-5">
+        <FormControl
+          label="Bench name"
+          type="text"
+          v-model="name"
+          placeholder="my-bench"
+          @input="error = ''"
+          @keyup.enter="createBench"
+        />
+        <div>
+          <span class="mb-1.5 block text-xs text-ink-gray-5">Process manager</span>
+          <div class="grid grid-cols-2 gap-2">
+            <button
+              v-for="opt in processManagerOptions"
+              :key="opt.value"
+              type="button"
+              class="rounded-lg border px-3 py-2 text-left transition-colors"
+              :class="processManager === opt.value
+                ? 'border-outline-gray-3 bg-surface-gray-2'
+                : 'border-outline-gray-2 hover:bg-surface-gray-1'"
+              @click="processManager = opt.value"
+            >
+              <span class="block text-sm font-medium text-ink-gray-9">{{ opt.label }}</span>
+              <span class="block text-xs text-ink-gray-5">{{ opt.hint }}</span>
+            </button>
+          </div>
+        </div>
+        <div>
+          <FormControl
+            label="Admin domain"
+            type="text"
+            v-model="adminDomain"
+            placeholder="my-admin.example.com"
+            @input="error = ''"
+            @keyup.enter="createBench"
+          />
+          <p class="mt-1.5 text-xs text-ink-gray-5">
+            The web address you'll use to open this bench.
+          </p>
+        </div>
+        <ErrorMessage v-if="error" :message="error" />
+        <p v-if="status" class="text-sm text-ink-gray-6">{{ status }}</p>
+      </div>
+    </template>
+    <template #actions>
+      <div class="flex justify-end gap-2">
+        <Button variant="ghost" @click="show = false">Cancel</Button>
+        <Button variant="solid" :loading="creating" @click="createBench">Create</Button>
+      </div>
+    </template>
+  </Dialog>
+</template>
